edit arithmetic in readfile function

+++ Plane.cpp
@@ -4,7 +4,7 @@
 #include "Plane.h"
 
 
-	Plane::Plane(glm::vec3 normal, glm::vec3 position, glm::vec3 ambient, glm::vec3 diffuse, glm::vec3 specular, int shininess) {
+	Plane::Plane(glm::vec3 normal, glm::vec3 position, glm::vec3 ambient, glm::vec3 diffuse, glm::vec3 specular, float shininess) {
 		nor = normal;
 		pos = position;
 		amb = ambient;
